+++ src/compas_timber/elements/__init__.py
@@ -3,10 +3,6 @@ from .plate import Plate
 from .panel import Panel
 from .fastener import Fastener
 from .fastener import FastenerTimberInterface
-from .features import BrepSubtraction
-from .features import CutFeature
-from .features import DrillFeature
-from .features import MillVolume
 from .fasteners.ball_node_fastener import BallNodeFastener
 from .fasteners.plate_fastener import PlateFastener
 from .plate_geometry import PlateGeometry
@@ -16,10 +12,6 @@ __all__ = [
     "Plate",
     "Fastener",
     "FastenerTimberInterface",
-    "CutFeature",
-    "DrillFeature",
-    "MillVolume",
-    "BrepSubtraction",
     "BallNodeFastener",
     "PlateFastener",
     "Panel",
