+++ tests/compas_timber/test_beam.py
@@ -509,3 +509,53 @@ def test_transform_invalidates_cached_timber_attributes(beam):
     assert new_blank.frame != original_blank.frame
     assert new_ref_frame.point != original_ref_frame.point
     assert not TOL.is_close(new_blank.frame.point.distance_to_point(original_blank.frame.point), 0.0)
+
+
+def test_add_extend_start_feature():
+    start = Point(0.0, 0.0, 0.0)
+    end = Point(1.0, 0.0, 0.0)
+    beam = Beam.from_endpoints(start, end, 0.06, 0.12)
+    beam.add_blank_extension(start=0.10, end=0.0)
+
+    assert close(beam.blank.xsize, 1.1)
+
+
+def test_add_extend_end_feature():
+    start = Point(0.0, 0.0, 0.0)
+    end = Point(1.0, 0.0, 0.0)
+    beam = Beam.from_endpoints(start, end, 0.06, 0.12)
+
+    beam.add_blank_extension(start=0.0, end=0.10)
+
+    assert close(beam.blank.xsize, 1.1)
+    # TODO: assert was extended at end and not at start
+
+
+def test_extend_both_start_end():
+    start = Point(0.0, 0.0, 0.0)
+    end = Point(1.0, 0.0, 0.0)
+    beam = Beam.from_endpoints(start, end, 0.06, 0.12)
+
+    beam.add_blank_extension(start=0.10, end=0.10)
+
+    assert close(beam.blank.xsize, 1.2)
+
+
+def test_accumulate_extension():
+    start = Point(0.0, 0.0, 0.0)
+    end = Point(1.0, 0.0, 0.0)
+    beam = Beam.from_endpoints(start, end, 0.06, 0.12)
+    beam.add_blank_extension(start=0.0, end=0.10)
+    beam.add_blank_extension(start=0.0, end=0.20)
+
+    # max extension is used
+    assert close(beam.blank.xsize, 1.20)
+
+
+def test_remove_parametric_extension():
+    start = Point(0.0, 0.0, 0.0)
+    end = Point(1.0, 0.0, 0.0)
+    beam = Beam.from_endpoints(start, end, 0.06, 0.12)
+    beam.add_blank_extension(start=0.0, end=0.10)
+
+    assert close(beam.blank.xsize, 1.10)
